requirements.txt
flask
pillow
opencv-python
pybase64
ultralytics